rework pocket service
