+++ src/services/pocket/List.tsx
@@ -1,24 +1,20 @@
 import { FC } from 'react';
 
 import { List as RawList, ListProps } from 'components/List';
-import { isConnected } from './helpers';
 import { PocketItem, itemToListItem } from './item';
-import { ConnectButton } from './ConnectButton';
 import { getQueryKeys } from './const';
 import { archiveItem, deleteItem } from './apiActions';
 import { get, search } from './apiGet';
 
 const PocketList = RawList as FC<ListProps<PocketItem>>;
 
-export const List: FC = () => {
-  return isConnected() ? (
-    <PocketList
-      getQueryKey={getQueryKeys}
-      get={get}
-      search={search}
-      archiveItem={archiveItem}
-      deleteItem={deleteItem}
-      itemToListItem={itemToListItem}
-    />
-  ) : <ConnectButton />;
-};
+export const List: FC = () => (
+  <PocketList
+    getQueryKey={getQueryKeys}
+    get={get}
+    search={search}
+    archiveItem={archiveItem}
+    deleteItem={deleteItem}
+    itemToListItem={itemToListItem}
+  />
+);
