+++ src/services/pocket/helpers.ts
@@ -4,34 +4,28 @@ import {
   setLocalStorageValue as set,
   deleteLocalStorageValue as del
 } from 'helpers/localstorage';
+import { LocalStorageKeys, localStorageKeyCodes } from './const';
 
-type Keys = 'code' | 'token' | 'username';
-const localStorageKeys: Record<Keys, string> = {
-  code: 'pocketCode',
-  token: 'pocketToken',
-  username: 'pocketUsername'
-};
-
-const getKeys = (): Keys[] => Object.keys(localStorageKeys) as Keys[];
+const getKeys = (): LocalStorageKeys[] => Object.keys(localStorageKeyCodes) as LocalStorageKeys[];
 
 export const isConnected = (): boolean => getKeys()
-  .map((key) => Boolean(get(localStorageKeys, key)))
+  .map((key) => Boolean(get(localStorageKeyCodes, key)))
   .reduce((a, b) => a && b);
 
-export const deleteAllKeys = () => getKeys().forEach((key) => del(localStorageKeys, key));
+export const deleteAllKeys = () => getKeys().forEach((key) => del(localStorageKeyCodes, key));
 
 export const getPocketKey = (): string | undefined => getEnvVar('pocketKey');
 export const getPocketRedirectUri = (): string =>
   chrome.extension.getURL('services/pocket/oauth.html');
 
-export const getPocketCode = () => get(localStorageKeys, 'code');
+export const getPocketCode = () => get(localStorageKeyCodes, 'code');
 export const setPocketCode = (code: string) =>
-  set(localStorageKeys, 'code', code);
+  set(localStorageKeyCodes, 'code', code);
 
-export const getPocketToken = () => get(localStorageKeys, 'token');
+export const getPocketToken = () => get(localStorageKeyCodes, 'token');
 export const setPocketToken = (token: string) =>
-  set(localStorageKeys, 'token', token);
+  set(localStorageKeyCodes, 'token', token);
 
-export const getPocketUsername = () => get(localStorageKeys, 'username');
+export const getPocketUsername = () => get(localStorageKeyCodes, 'username');
 export const setPocketUsername = (username: string) =>
-  set(localStorageKeys, 'username', username);
+  set(localStorageKeyCodes, 'username', username);
